Lec13: the MPY is almost done. I've stalled on the MSP430 example:
nowhere is given an example of why there should be a NOP when using
indirect addressing mode with the MPY. So, I had to look up the
instructions' cycles and do the math. I believe, so far, my 2 examples
for the MSP430 are unique in the Internet.

Time for beer, it's Friday, fellas ...