Lec6: corrected the up/down counter traces to reset on 512, and not on 511.
Added a 5V supply on the circuit diagram of the inverter to make it more
believable (due to the high-power low-side MOSFET).

Corrected the password - for the watchdog of the MSP430 the pass is 0x5A
but it is in the high-byte so the mask should be 0x5A00, and not 0x5A.

I'm mentioning the start-up timer but I have it in the next lecture. I've
put a "pointer" to the next lecture for the curious ones.